Update template dependencies, switch license
Switch to an even more permissive license to enable more projects
to use the template without author attribution and a binding license.

General housekeeping (see changelog)

Major dependecy updates with breaking changes. Consult the changelogs
listed for the updated dependencies (see changelog) when updating
a project based on an older version of this template.